+++ src/constructors/constructWithOptions.js
@@ -15,6 +15,13 @@ export default function constructWithOptions(
     throw new StyledError(1, String(tag));
   }
 
+  // Creates resolver function which wraps the logic for resolving component attrs.
+  // Currently it supports both ways for defining component attrs:
+  // - plain object -> styled.input.attrs({ example: 'value' })
+  // - factory function -> styled.input.attrs(props => ({ example: props.value })
+  const resolveAttr = (context, attrsObjOrFn) =>
+    typeof attrsObjOrFn === 'function' ? attrsObjOrFn(context) : attrsObjOrFn;
+
   /* This is callable directly as a template function */
   // $FlowFixMe: Not typed to avoid destructuring arguments
   const templateFunction = (...args) => componentConstructor(tag, options, css(...args));
@@ -22,10 +29,13 @@ export default function constructWithOptions(
   /* If config methods are called, wrap up a new template function and merge options */
   templateFunction.withConfig = config =>
     constructWithOptions(componentConstructor, tag, { ...options, ...config });
+
+  /* attrs could be either a plain object or a function (props => ({ attrs })) */
   templateFunction.attrs = attrs =>
     constructWithOptions(componentConstructor, tag, {
       ...options,
       attrs: { ...(options.attrs || EMPTY_OBJECT), ...attrs },
+      resolveAttr,
     });
 
   return templateFunction;